move download page button to reader settings drawer
