+++ NEWS.md
@@ -13,6 +13,7 @@ border-radius: 128px;
   - Allow filtering by None values when None values exist.
   - Show only usable filters for current view as filter choices.
   - PDF viewer shows loading, rendering and password errors.
+  - Move download page button to the reader settings drawer.
 - Fixes
   - Filters with compound names were not loading choices.
   - Handle an iOS bug with downloading pages and comics inside a PWA.
