+++ frontend/src/stores/reader.js
@@ -4,6 +4,7 @@ import BROWSER_API from "@/api/v3/browser";
 import API from "@/api/v3/reader";
 import CHOICES from "@/choices";
 import router from "@/router";
+import { getFullComicName } from "@/comic-name";
 
 const NULL_READER_SETTINGS = {
   // Must be null so axios doesn't throw them out when sending.
@@ -111,6 +112,12 @@ export const useReaderStore = defineStore("reader", {
       }
       return classes;
     },
+    title(state) {
+       if (state.comic) {
+          return getFullComicName(state.comic);
+      }
+      return "";
+    }
   },
   actions: {
     ///////////////////////////////////////////////////////////////////////////
